Move logic related to changes in vocabs to custom useVocabs hook
