+++ src/index.js
@@ -1,6 +1,5 @@
 import React from 'react'
 import ReactDOM from 'react-dom'
-import { BrowserRouter as Router } from 'react-router-dom'
 import App from './App'
 import GlobalStyles from './common/GlobalStyles'
 import * as serviceWorker from './serviceWorker'
@@ -8,9 +7,7 @@ import * as serviceWorker from './serviceWorker'
 ReactDOM.render(
   <>
     <GlobalStyles />
-    <Router>
-      <App />
-    </Router>
+    <App />
   </>,
   document.getElementById('root')
 )
